+++ introspect/enums_test.py
@@ -41,10 +41,11 @@ class EnumsTest(absltest.TestCase):
         tuple(enum_decl.values.items()), (('mjENBL_OVERRIDE', 1<<0),
                                           ('mjENBL_ENERGY', 1<<1),
                                           ('mjENBL_FWDINV', 1<<2),
-                                          ('mjENBL_SENSORNOISE', 1<<3),
-                                          ('mjENBL_MULTICCD', 1<<4),
-                                          ('mjENBL_ISLAND', 1<<5),
-                                          ('mjNENABLE', 6)))
+                                          ('mjENBL_INVDISCRETE', 1<<3),
+                                          ('mjENBL_SENSORNOISE', 1<<4),
+                                          ('mjENBL_MULTICCD', 1<<5),
+                                          ('mjENBL_ISLAND', 1<<6),
+                                          ('mjNENABLE', 7)))
 
   # values mostly increment by one with occasional overrides
   def test_mjtGeom(self):  # pylint: disable=invalid-name